requirements.txt
ultralytics>=8.0.0
pyyaml>=6.0
opencv-python>=4.8.0
numpy>=1.24
onnx>=1.14
onnxruntime>=1.15
onnxslim>=0.1.34
